+++ scanpipe/management/commands/list-pipelines.py
@@ -28,13 +28,32 @@ scanpipe_app = apps.get_app_config("scanpipe")
 
 
 class Command(BaseCommand):
-    help = "Lists available pipelines."
+    help = (
+        "Displays a list of available pipelines. "
+        "Use --verbosity=2 to include details of each pipeline's steps."
+    )
 
     def handle(self, *args, **options):
         verbosity = options["verbosity"]
 
         for module_name, pipeline_class in scanpipe_app.pipelines.items():
-            self.stdout.write(module_name)
-            if verbosity > 1:
-                pipeline_info = pipeline_class.get_info()
-                self.stdout.write(pipeline_info.get("summary"))
+            msg = self.style.HTTP_INFO(module_name)
+            if pipeline_class.is_addon:
+                msg += " (addon)"
+            self.stdout.write(msg)
+            pipeline_info = pipeline_class.get_info()
+            if verbosity >= 1:
+                self.stdout.write(pipeline_info["summary"])
+
+            if verbosity >= 2:
+                steps = pipeline_info["steps"]
+                for step_info in steps:
+                    step_name = step_info["name"]
+                    step_doc = step_info["doc"].replace("\n", "\n   ")
+                    step_groups = step_info["groups"]
+                    self.stdout.write(f" > [{step_name}]: {step_doc}")
+                    if step_groups:
+                        self.stdout.write(f"   {{Group}}: {','.join(step_groups)}")
+
+            if verbosity >= 1:
+                self.stdout.write("\n")